requirements.txt
asgiref==3.6.0
Django==4.2
sqlparse==0.4.3
djangorestframework==3.14.0
pytz==2023.3
